+++ src/main.js
@@ -4,10 +4,7 @@ import getStream from 'get-stream'
 // Fetch all available Node versions by making a HTTP request to Node website
 // Versions are already sorted from newest to oldest
 const allNodeVersions = async function(opts) {
-  const response = await fetchNodeWebsite(INDEX_PATH, {
-    progress: false,
-    ...opts,
-  })
+  const response = await fetchNodeWebsite(INDEX_PATH, opts)
   const content = await getStream(response)
   const index = JSON.parse(content)
   const versions = index.map(getVersionField)